+++ apps/endringsmelding-pleiepenger/src/app/api/useSøknadInitialData.ts
@@ -6,7 +6,7 @@ import { APP_VERSJON } from '../constants/APP_VERSJON';
 import { SøknadRoutes } from '../søknad/config/SøknadRoutes';
 import { Arbeidsgiver } from '../types/Arbeidsgiver';
 import { IngenTilgangÅrsak } from '../types/IngenTilgangÅrsak';
-import { K9Sak } from '../types/K9Sak';
+import { isK9Sak, isUgyldigK9SakFormat, K9Sak, UgyldigK9SakFormat } from '../types/K9Sak';
 import { RequestStatus } from '../types/RequestStatus';
 import { Søker } from '../types/Søker';
 import { SøknadContextState } from '../types/SøknadContextState';
@@ -14,6 +14,7 @@ import { TimerEllerProsent } from '../types/TimerEllerProsent';
 import appSentryLogger from '../utils/appSentryLogger';
 import { getEndringsdato, getMaksEndringsperiode } from '../utils/endringsperiode';
 import { getSakFromK9Sak } from '../utils/getSakFromK9Sak';
+import { getSakOgArbeidsgivereDebugInfo, maskK9Sak } from '../utils/getSakOgArbeidsgivereDebugInfo';
 import { getPeriodeForArbeidsgiverOppslag, getSamletDateRangeForK9Saker } from '../utils/k9SakUtils';
 import { tilgangskontroll } from '../utils/tilgangskontroll';
 import { arbeidsgivereEndpoint } from './endpoints/arbeidsgivereEndpoint';
@@ -103,7 +104,12 @@ const setupSøknadInitialData = async (
             return getSakFromK9Sak(persistedSak, arbeidsgivere, endringsperiode);
         }
         if (k9saker.length === 1) {
-            return getSakFromK9Sak(k9saker[0], arbeidsgivere, endringsperiode);
+            const sak = getSakFromK9Sak(k9saker[0], arbeidsgivere, endringsperiode);
+            appSentryLogger.logInfo(
+                'debug.maskedSakInfo',
+                JSON.stringify(getSakOgArbeidsgivereDebugInfo(k9saker[0], sak, arbeidsgivere, endringsperiode))
+            );
+            return sak;
         }
         return undefined;
     };
@@ -130,11 +136,25 @@ function useSøknadInitialData(): SøknadInitialDataState {
         try {
             const maksEndringsperiode = getMaksEndringsperiode(getEndringsdato());
 
-            const [søker, k9saker, lagretSøknadState] = await Promise.all([
+            const [søker, k9sakerResult, lagretSøknadState] = await Promise.all([
                 søkerEndpoint.fetch(),
                 sakerEndpoint.fetch(),
                 søknadStateEndpoint.fetch(),
             ]);
+
+            const ugyldigk9FormatSaker: UgyldigK9SakFormat[] = k9sakerResult.filter(isUgyldigK9SakFormat);
+            const k9saker: K9Sak[] = k9sakerResult.filter(isK9Sak);
+
+            /** Dersom vi ikke klarer å parse saken */
+            if (k9sakerResult.length === 1 && ugyldigk9FormatSaker.length === 1) {
+                setInitialData({
+                    status: RequestStatus.success,
+                    kanBrukeSøknad: false,
+                    årsak: IngenTilgangÅrsak.harUgyldigK9FormatSak,
+                    søker,
+                });
+                return Promise.resolve();
+            }
 
             const samletTidsperiode = getSamletDateRangeForK9Saker(k9saker);
 
@@ -163,6 +183,23 @@ function useSøknadInitialData(): SøknadInitialDataState {
                     årsak: resultat.årsak,
                     søker,
                 });
+                if (k9saker.length === 1) {
+                    appSentryLogger.logInfo(
+                        'IkkeTilgangSakInfo',
+                        JSON.stringify({
+                            årsak: resultat.årsak,
+                            sak: maskK9Sak(k9saker[0]),
+                        })
+                    );
+                }
+                if (k9saker.length > 1) {
+                    appSentryLogger.logInfo(
+                        'IkkeTilgangSakInfo',
+                        JSON.stringify({
+                            årsak: resultat.årsak,
+                        })
+                    );
+                }
                 return Promise.resolve();
             }
 
@@ -185,7 +222,6 @@ function useSøknadInitialData(): SøknadInitialDataState {
                     status: RequestStatus.forbidden,
                 });
             } else {
-                appSentryLogger.logError('fetchInitialData', error);
                 setInitialData({
                     status: RequestStatus.error,
                     error,