+++ apps/omsorgspengesoknad/src/app/utils/appSentryLogger.ts
@@ -1,5 +1,5 @@
 import getSentryLoggerForApp from '@navikt/sif-common-sentry';
 
-const appSentryLogger = getSentryLoggerForApp('omsorgspengesøknad', ['omsorgspengesoknad.nav.no/dist/js'], [/401/]);
+const appSentryLogger = getSentryLoggerForApp('omsorgspengesøknad', ['omsorgspengesoknad.nav.no/dist/js']);
 
 export default appSentryLogger;