+++ apps/sif-ettersending/e2e/cypress/support/commands.ts
@@ -37,3 +37,9 @@
 // }
 
 import 'cypress-file-upload';
+
+Cypress.on('uncaught:exception', (err, runnable) => {
+    // returning false here prevents Cypress from
+    // failing the test
+    return false;
+});