+++ apps/sif-ettersending/src/app/hooks/useSoknadEssentials.ts
@@ -1,5 +1,7 @@
 import { useState } from 'react';
 import useEffectOnce from '@navikt/sif-common-core-ds/lib/hooks/useEffectOnce';
+import { isForbidden, isUnauthorized } from '@navikt/sif-common-core-ds/lib/utils/apiUtils';
+import { isObjectLike } from 'lodash';
 import getSokerRemoteData from '../api/getSoker';
 import getSoknadTempStorage from '../api/getSoknadTempStorage';
 import { ApplicationType } from '../types/ApplicationType';
@@ -7,7 +9,6 @@ import { Person } from '../types/Person';
 import { RequestStatus } from '../types/RequestStatus';
 import { SoknadTempStorageData } from '../types/SoknadTempStorageData';
 import appSentryLogger from '../utils/appSentryLogger';
-import { isForbidden, isUnauthorized } from '@navikt/sif-common-core-ds/lib/utils/apiUtils';
 import { navigateToLoginPage } from '../utils/navigationUtils';
 
 export type SoknadEssentials = { søker: Person; mellomlagring?: SoknadTempStorageData };
@@ -15,6 +16,10 @@ export type SoknadEssentials = { søker: Person; mellomlagring?: SoknadTempStora
 type SøknadInitialSuccess = {
     status: RequestStatus.success;
     data: SoknadEssentials;
+};
+
+type SøknadInitialIkkeTilgang = {
+    status: RequestStatus.ikkeTilgang;
 };
 
 type SøknadInitialFailed = {
@@ -26,7 +31,25 @@ type SøknadInitialLoading = {
     status: RequestStatus.loading;
 };
 
-export type SøknadInitialDataState = SøknadInitialSuccess | SøknadInitialFailed | SøknadInitialLoading;
+export type SøknadInitialDataState =
+    | SøknadInitialSuccess
+    | SøknadInitialFailed
+    | SøknadInitialLoading
+    | SøknadInitialIkkeTilgang;
+
+const isUnknownAxiosError = (error: any) => {
+    try {
+        return (
+            error !== undefined &&
+            isObjectLike(error) &&
+            error.code === 'ERR_NETWORK' &&
+            error.status === 'None' &&
+            error.number[0] === undefined
+        );
+    } catch (e) {
+        return false;
+    }
+};
 
 function useSoknadEssentials(søknadstype: ApplicationType): SøknadInitialDataState {
     const [initialData, setInitialData] = useState<SøknadInitialDataState>({ status: RequestStatus.loading });
@@ -40,11 +63,17 @@ function useSoknadEssentials(søknadstype: ApplicationType): SøknadInitialDataS
             });
             return Promise.resolve();
         } catch (error) {
-            if (isForbidden(error) || isUnauthorized(error)) {
+            if (isUnauthorized(error)) {
                 navigateToLoginPage(søknadstype);
                 return Promise.reject(error);
+            } else if (isForbidden(error)) {
+                setInitialData({
+                    status: RequestStatus.ikkeTilgang,
+                });
             } else {
-                appSentryLogger.logError('fetchInitialData', error);
+                if (!isUnknownAxiosError(error)) {
+                    appSentryLogger.logError('fetchInitialData', error);
+                }
                 setInitialData({
                     status: RequestStatus.error,
                     error,