+++ apps/sif-ettersending/src/app/utils/appSentryLogger.ts
@@ -1,5 +1,5 @@
 import getSentryLoggerForApp from '@navikt/sif-common-sentry';
 
-const appSentryLogger = getSentryLoggerForApp('ettersending', ['sykdom-i-familien'], [/401/]);
+const appSentryLogger = getSentryLoggerForApp('ettersending', ['sykdom-i-familien']);
 
 export default appSentryLogger;