+++ packages/sif-common-soknad-ds/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @navikt/sif-common-soknad-ds
 
+## 6.0.1
+
+### Patch Changes
+
+-   Updated dependencies [7fb116b5]
+    -   @navikt/sif-common-sentry@0.9.0
+
+## 6.0.1
+
+### Patch Changes
+
+-   7c6db10: Pakkeoppdateringer
+-   Updated dependencies [7c6db10]
+    -   @navikt/sif-common-formik-ds@0.11.2
+    -   @navikt/appstatus-react-ds@2.0.1
+
 ## 6.0.0
 
 ### Patch Changes